Assignment-42labels: Revised

+++ app/controllers/comments_controller.rb
@@ -5,51 +5,59 @@ class CommentsController < ApplicationController
 
 
   def create
-     @post = Post.find_by(id: params[:post_id])
-     @topic = Topic.find_by(id: params[:topic_id])
+    #  @post = Post.find_by(id: params[:post_id])
+    #  @topic = Topic.find_by(id: params[:topic_id])
 
-     if @post
+     if params[:post_id]
+       @post = Post.find_by(id: params[:post_id])
        comment = @post.comments.new(comment_params)
+       comment.user = current_user
+
+       if comment.save
+         flash[:notice] = "Comment saved successfully."
+         redirect_to [@post.topic, @post]
+       else
+         flash[:alert] = "Comment was unable to be saved. Try harder."
+         redirect_to [@post.topic, @post]
+       end
      else
+       @topic = Topic.find_by(id: params[:topic_id])
        comment = @topic.comments.new(comment_params)
-     end
+       comment.user = current_user
 
-     comment.user = current_user
-
-     if comment.save
-       flash[:notice] = "Comment saved successfully."
-     else
-       flash[:alert] = "Comment was unable to be saved. Try harder."
-     end
-
-     if @post
-       redirect_to [@post.topic, @post]
-     else
-       redirect_to @topic
+       if comment.save
+         flash[:notice] = "Comment saved successfully."
+         redirect_to @topic
+       else
+         flash[:alert] = "Comment was unable to be saved. Try harder."
+         redirect_to @topic
+       end
      end
   end
 
   def destroy
-     @post = Post.find_by(id: params[:post_id])
-     @topic = Topic.find_by(id: params[:topic_id])
+     if params[:post_id]
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.find(params[:id])
 
-     if @post
-       comment = @post.comments.find_by(id: params[:id])
-     else
-       comment = @topic.comments.find_by(id: params[:id])
-     end
-
-     if comment.destroy
-       flash[:notice] = "Comment was deleted successfully."
-     else
-       flash[:alert] = "Comment couldn't be deleted. Try again."
-     end
-
-     if @post
-       redirect_to [@post.topic, @post]
-     else
-       redirect_to @topic
-     end
+        if comment.destroy
+          flash[:notice] = "Comment was deleted."
+          redirect_to [@post.topic, @post]
+        else
+          flash[:alert] = "Comment couldn't be deleted. Try again."
+          redirect_to [@post.topic, @post]
+        end
+      elsif params[:topic_id]
+        @topic = Topic.find(params[:topic_id])
+        comment = @topic.comments.find(params[:id])
+        if comment.destroy
+          flash[:notice] = "Comment was deleted."
+          redirect_to @topic
+        else
+          flash[:notice] = "Comment couldn't be deleted."
+          redirect_to @topic
+        end
+      end
    end
 
 
@@ -60,10 +68,14 @@ class CommentsController < ApplicationController
   end
 
   def authorize_user
-     comment = Comment.find(params[:id])
-     unless current_user == comment.user || current_user.admin?
-       flash[:alert] = "You do not have permission to delete a comment."
-       redirect_to [comment.post.topic, comment.post]
-     end
+    comment = Comment.find(params[:id])
+    unless current_user == comment.user || current_user.admin?
+      flash[:alert] = "You do not have permission to delete a comment."
+      if params[:post_id]
+        redirect_to [comment.post.topic, comment.post]
+      elsif params[:topic_id]
+        redirect_to comment.topic
+      end
+    end
   end
 end
